move aliasing stuff into its own module
