+++ cogs5e/sheetManager.py
@@ -12,7 +12,7 @@ from discord.ext import commands
 from discord.ext.commands.cooldowns import BucketType
 
 from cogs5e.funcs import attackutils, checkutils, targetutils
-from cogs5e.funcs.scripting import helpers
+from aliasing import helpers
 from cogs5e.models.character import Character
 from cogs5e.models.embeds import EmbedWithCharacter
 from cogs5e.models.errors import ExternalImportError
